feat(website, ui): add admin panel boilerplate
