+++ src/app/layout.tsx
@@ -1,9 +1,12 @@
 import { Montserrat } from "next/font/google";
 import type { Metadata } from "next";
 import type { ReactNode, FC } from "react";
-import { TRPCProvider } from "@khenzii-dev/providers";
-import style from "@khenzii-dev/styles/layout.module.scss";
 import clsx from "clsx";
+import {
+    TRPCProvider,
+    SessionProviderWrapper,
+} from "@khenzii-dev/providers";
+import style from "@khenzii-dev/styles/layout.module.scss";
 
 const montserrat = Montserrat({
     weight: "600",
@@ -22,11 +25,13 @@ export const metadata: Metadata = {
 
 const RootLayout: FC<{ children: ReactNode }> = ({ children }) => (
     <TRPCProvider>
-        <html lang="en">
-            <body className={clsx(style.layout, montserrat.className)}>
-                {children}
-            </body>
-        </html>
+        <SessionProviderWrapper>
+            <html lang="en">
+                <body className={clsx(style.layout, montserrat.className)}>
+                    {children}
+                </body>
+            </html>
+        </SessionProviderWrapper>
     </TRPCProvider>
 );
