+++ src/providers/index.ts
@@ -1,1 +1,2 @@
 export * from "./trpc_provider";
+export * from "./session_provider_wrapper";
