+++ src/ui/organisms/admin_login/index.stories.tsx
@@ -1,0 +1,12 @@
+import { AdminLogin } from "@khenzii-dev/ui/organisms";
+import type { StoriesType } from "@khenzii-dev/ui/types";
+
+export default {
+    component: AdminLogin,
+    title: 'AdminLogin',
+    tags: ['autodocs'],
+};
+
+export const Default: StoriesType<Record<string, null>> = {
+    args: {},
+};
