+++ src/ui/organisms/index.ts
@@ -1,3 +1,5 @@
 export * from "./footer";
 export * from "./nav";
 export * from "./projects";
+export * from "./admin_login";
+export * from "./protected_page";
